chore: refactor test
