+++ tests/fixtures/rules/infinite-reactive-loop/invalid/await/test01-input.svelte
@@ -1,0 +1,9 @@
+<script>
+  let a = 0
+
+  $: (async () => {
+    a = a + 1
+    await doSomething()
+    a = a + 1
+  })()
+</script>
