+++ tests/fixtures/rules/infinite-reactive-loop/invalid/function-call/test01-errors.yaml
