+++ tests/fixtures/rules/infinite-reactive-loop/invalid/function-call/test02-errors.yaml
