+++ tests/fixtures/rules/infinite-reactive-loop/invalid/function-call/test03-errors.yaml
