+++ tests/fixtures/rules/infinite-reactive-loop/invalid/function-call/test03-input.svelte
@@ -3,14 +3,18 @@
   const fetch = async () => {
     await new Promise((resolve) => setTimeout(resolve, 100))
   }
-  const doSomething = () => {
-    fetch().then(() => {
-      a += 1
-    })
+  const doSomething = async () => {
+    await fetch()
+    a += 1
   }
 
   $: (async () => {
     console.log(a)
+    await doSomething()
+  })()
+
+  $: (async () => {
+    // should not report here
     await doSomething()
   })()
 </script>
