+++ tests/fixtures/rules/infinite-reactive-loop/invalid/function-call/test04-errors.yaml
