+++ tests/fixtures/rules/infinite-reactive-loop/invalid/function-call/test05-input.svelte
