+++ tests/fixtures/rules/infinite-reactive-loop/invalid/function-call/test06-input.svelte
