+++ tests/fixtures/rules/infinite-reactive-loop/invalid/function-call/test07-errors.yaml
