+++ tests/fixtures/rules/infinite-reactive-loop/invalid/function-call/test08-errors.yaml
@@ -1,5 +1,5 @@
 - message: Possibly it may occur an infinite reactive loop.
-  line: 7
+  line: 6
   column: 5
   suggestions: null
 - message: Possibly it may occur an infinite reactive loop because this function may update `foo`.
