+++ tests/fixtures/rules/infinite-reactive-loop/invalid/function-call/test08-input.svelte
@@ -1,5 +1,4 @@
 <script>
-  let obj = { a: 0 }
   let foo = { obj: 1 }
 
   const doSomething = async () => {
@@ -8,8 +7,10 @@
   }
 
   $: (async () => {
+    const obj = { a: 0 }
     console.log(obj)
     await doSomething()
     foo.obj += 1
+    obj.a = 1
   })()
 </script>
