+++ tests/fixtures/rules/infinite-reactive-loop/invalid/function-call/test09-input.svelte
