+++ tests/fixtures/rules/infinite-reactive-loop/invalid/function-call/test10-errors.yaml
@@ -1,0 +1,4 @@
+- message: Possibly it may occur an infinite reactive loop.
+  line: 9
+  column: 30
+  suggestions: null
