+++ tests/fixtures/rules/infinite-reactive-loop/invalid/function-call/test10-input.svelte
@@ -1,0 +1,11 @@
+<script>
+  let obj = { a: 0 }
+
+  const doSomething = (a, b) => {
+    console.log({ a, b })
+  }
+
+  $: (async () => {
+    doSomething((await "a", (obj.a += 1)))
+  })()
+</script>
