+++ tests/fixtures/rules/infinite-reactive-loop/invalid/promise/test01-errors.yaml
@@ -1,12 +1,12 @@
 - message: Possibly it may occur an infinite reactive loop.
-  line: 5
-  column: 5
+  line: 6
+  column: 7
   suggestions: null
-- message: Possibly it may occur an infinite reactive loop because this function may update `obj`.
-  line: 10
-  column: 11
+- message: Possibly it may occur an infinite reactive loop.
+  line: 7
+  column: 7
   suggestions: null
 - message: Possibly it may occur an infinite reactive loop.
   line: 11
-  column: 5
+  column: 7
   suggestions: null
