+++ tests/fixtures/rules/infinite-reactive-loop/invalid/promise/test01-input.svelte
@@ -1,0 +1,14 @@
+<script>
+  let a = 0
+
+  $: (() => {
+    Promise.resolve().then(() => {
+      a = a + 1
+      a += 1
+    })
+
+    Promise.resolve().catch(() => {
+      a = a + 1
+    })
+  })()
+</script>
