+++ tests/fixtures/rules/infinite-reactive-loop/invalid/promise/test02-errors.yaml
@@ -1,0 +1,8 @@
+- message: Possibly it may occur an infinite reactive loop.
+  line: 10
+  column: 9
+  suggestions: null
+- message: Possibly it may occur an infinite reactive loop.
+  line: 13
+  column: 9
+  suggestions: null
