+++ tests/fixtures/rules/infinite-reactive-loop/invalid/promise/test02-input.svelte
@@ -1,0 +1,17 @@
+<script>
+  let a = 0
+
+  $: {
+    a = a + 1
+    new Promise((resolve, reject) => {
+      /** do something */
+    })
+      .then(() => {
+        a = a + 1
+      })
+      .catch(() => {
+        a = a + 1
+      })
+    a = a + 1
+  }
+</script>
