+++ tests/fixtures/rules/infinite-reactive-loop/invalid/promise/test03-errors.yaml
@@ -1,12 +1,8 @@
 - message: Possibly it may occur an infinite reactive loop.
-  line: 7
-  column: 7
+  line: 8
+  column: 9
   suggestions: null
 - message: Possibly it may occur an infinite reactive loop.
-  line: 10
-  column: 7
-  suggestions: null
-- message: Possibly it may occur an infinite reactive loop.
-  line: 13
-  column: 7
+  line: 11
+  column: 9
   suggestions: null
