+++ tests/fixtures/rules/infinite-reactive-loop/invalid/promise/test04-errors.yaml
@@ -1,0 +1,4 @@
+- message: Possibly it may occur an infinite reactive loop.
+  line: 7
+  column: 5
+  suggestions: null
