+++ tests/fixtures/rules/infinite-reactive-loop/invalid/promise/test04-input.svelte
@@ -1,0 +1,9 @@
+<script>
+  let a = 0
+
+  $: (async () => {
+    a = a + 1
+    await Promise.resolve()
+    a = a + 1
+  })()
+</script>
