+++ tests/fixtures/rules/infinite-reactive-loop/invalid/promise/test05-errors.yaml
@@ -1,8 +1,8 @@
 - message: Possibly it may occur an infinite reactive loop.
-  line: 12
-  column: 24
+  line: 7
+  column: 9
   suggestions: null
 - message: Possibly it may occur an infinite reactive loop.
-  line: 13
-  column: 25
+  line: 10
+  column: 9
   suggestions: null
