+++ tests/fixtures/rules/infinite-reactive-loop/invalid/promise/test05-input.svelte
@@ -1,0 +1,13 @@
+<script>
+  let a = 0
+
+  $: (() => {
+    Promise.resolve()
+      .catch(() => {
+        a = a + 1
+      })
+      .catch(() => {
+        a = a + 1
+      })
+  })()
+</script>
