+++ tests/fixtures/rules/infinite-reactive-loop/invalid/queueMicrotask/test01-errors.yaml
@@ -1,0 +1,8 @@
+- message: Possibly it may occur an infinite reactive loop.
+  line: 7
+  column: 7
+  suggestions: null
+- message: Possibly it may occur an infinite reactive loop.
+  line: 13
+  column: 7
+  suggestions: null
