+++ tests/fixtures/rules/infinite-reactive-loop/invalid/queueMicrotask/test01-input.svelte
@@ -1,0 +1,16 @@
+<script>
+  const queueMicrotask2 = queueMicrotask
+  let a = 0
+
+  $: {
+    queueMicrotask(() => {
+      a = a + 1
+    })
+  }
+
+  $: {
+    queueMicrotask2(() => {
+      a = a + 1
+    })
+  }
+</script>
