+++ tests/fixtures/rules/infinite-reactive-loop/invalid/setInterval/test01-input.svelte
@@ -1,0 +1,16 @@
+<script>
+  const setInterval2 = setInterval
+  let a = 0
+
+  $: {
+    setInterval(() => {
+      a = a + 1
+    })
+  }
+
+  $: {
+    setInterval2(() => {
+      a = a + 1
+    })
+  }
+</script>
