+++ tests/fixtures/rules/infinite-reactive-loop/invalid/setTimeout/test01-input.svelte
@@ -1,0 +1,24 @@
+<script>
+  const setTimeout2 = setTimeout
+  let a = 0
+
+  function doSomething(fn) {
+    fn()
+  }
+
+  $: {
+    setTimeout(() => {
+      a = a + 1
+    })
+  }
+
+  $: {
+    setTimeout2(() => {
+      a = a + 1
+    })
+
+    doSomething(() => {
+      a = a + 1
+    })
+  }
+</script>
