+++ tests/fixtures/rules/infinite-reactive-loop/invalid/setTimeout/test02-errors.yaml
