--- tests/fixtures/rules/infinite-reactive-loop/invalid/test01-errors.yaml
@@ -1,80 +1,0 @@
-- message: Possibly it may occur an infinite reactive loop.
-  line: 11
-  column: 7
-  suggestions: null
-- message: Possibly it may occur an infinite reactive loop.
-  line: 12
-  column: 7
-  suggestions: null
-- message: Possibly it may occur an infinite reactive loop.
-  line: 16
-  column: 7
-  suggestions: null
-- message: Possibly it may occur an infinite reactive loop.
-  line: 25
-  column: 9
-  suggestions: null
-- message: Possibly it may occur an infinite reactive loop.
-  line: 28
-  column: 9
-  suggestions: null
-- message: Possibly it may occur an infinite reactive loop.
-  line: 36
-  column: 9
-  suggestions: null
-- message: Possibly it may occur an infinite reactive loop.
-  line: 39
-  column: 9
-  suggestions: null
-- message: Possibly it may occur an infinite reactive loop.
-  line: 47
-  column: 5
-  suggestions: null
-- message: Possibly it may occur an infinite reactive loop.
-  line: 53
-  column: 5
-  suggestions: null
-- message: Possibly it may occur an infinite reactive loop.
-  line: 59
-  column: 5
-  suggestions: null
-- message: Possibly it may occur an infinite reactive loop.
-  line: 64
-  column: 7
-  suggestions: null
-- message: Possibly it may occur an infinite reactive loop.
-  line: 70
-  column: 7
-  suggestions: null
-- message: Possibly it may occur an infinite reactive loop.
-  line: 76
-  column: 7
-  suggestions: null
-- message: Possibly it may occur an infinite reactive loop.
-  line: 82
-  column: 7
-  suggestions: null
-- message: Possibly it may occur an infinite reactive loop.
-  line: 88
-  column: 7
-  suggestions: null
-- message: Possibly it may occur an infinite reactive loop.
-  line: 94
-  column: 7
-  suggestions: null
-- message: Possibly it may occur an infinite reactive loop.
-  line: 100
-  column: 7
-  suggestions: null
-- message: Possibly it may occur an infinite reactive loop.
-  line: 106
-  column: 7
-  suggestions: null
-- message: Possibly it may occur an infinite reactive loop.
-  line: 112
-  column: 5
-  suggestions: null
-- message: Possibly it may occur an infinite reactive loop.
-  line: 117
-  column: 5
-  suggestions: null
