--- tests/fixtures/rules/infinite-reactive-loop/invalid/test02-errors.yaml
@@ -1,40 +1,0 @@
-- message: Possibly it may occur an infinite reactive loop.
-  line: 7
-  column: 9
-  suggestions: null
-- message: Possibly it may occur an infinite reactive loop.
-  line: 8
-  column: 9
-  suggestions: null
-- message: Possibly it may occur an infinite reactive loop.
-  line: 11
-  column: 9
-  suggestions: null
-- message: Possibly it may occur an infinite reactive loop.
-  line: 12
-  column: 9
-  suggestions: null
-- message: Possibly it may occur an infinite reactive loop.
-  line: 17
-  column: 9
-  suggestions: null
-- message: Possibly it may occur an infinite reactive loop.
-  line: 20
-  column: 9
-  suggestions: null
-- message: Possibly it may occur an infinite reactive loop.
-  line: 25
-  column: 9
-  suggestions: null
-- message: Possibly it may occur an infinite reactive loop.
-  line: 28
-  column: 9
-  suggestions: null
-- message: Possibly it may occur an infinite reactive loop.
-  line: 33
-  column: 9
-  suggestions: null
-- message: Possibly it may occur an infinite reactive loop.
-  line: 36
-  column: 9
-  suggestions: null
