--- tests/fixtures/rules/infinite-reactive-loop/invalid/test06-input.svelte
@@ -1,15 +1,0 @@
-<script>
-  let a = 0
-
-  const func = () => {
-    return {
-      then: (fun) => fun(),
-      catch: (fun) => fun(),
-    }
-  }
-
-  $: (() => {
-    func().then(() => (a += 1))
-    func().catch(() => (a += 1))
-  })()
-</script>
