--- tests/fixtures/rules/infinite-reactive-loop/invalid/test09-input.svelte
@@ -1,16 +1,0 @@
-<script>
-  let a = 0
-
-  $: {
-    console.log(a)
-    setTimeout(() => {
-      a += 1
-    })
-    setInterval(() => {
-      a += 1
-    })
-    queueMicrotask(() => {
-      a += 1
-    })
-  }
-</script>
