--- tests/fixtures/rules/infinite-reactive-loop/invalid/test14-input.svelte
@@ -1,14 +1,0 @@
-<script>
-  let a = 0
-
-  const fetch = () => new Promise((resolve) => setTimeout(resolve, 1000))
-
-  const doSomething = async () => {
-    await fetch()
-  }
-
-  $: (async () => {
-    await doSomething()
-    a += 1
-  })()
-</script>
