+++ tests/fixtures/rules/infinite-reactive-loop/invalid/tick/test02-errors.yaml
@@ -1,0 +1,12 @@
+- message: Possibly it may occur an infinite reactive loop.
+  line: 17
+  column: 7
+  suggestions: null
+- message: Possibly it may occur an infinite reactive loop.
+  line: 23
+  column: 5
+  suggestions: null
+- message: Possibly it may occur an infinite reactive loop.
+  line: 28
+  column: 5
+  suggestions: null
